Create Shutdown.bat
copy/paste this code into a new notepad file - click the 'file' menu drop down and click "save as" - click the bottom right dropdown of file explorer and change filetype to all - name the file and add ".bat" to the end of the name

That is you're done. You can also change the Icon or make a shortcut.